vision server: qwen3-omni to extra; cheaper context-length launch; drop nightly-covered glm41v
